Clean up script first pass
The time and joule conversion factors are still hard coded.
This is because the bash code for time is not working and origin of joules factor is unknown.

There appears to be an issue with run-experiment where inputs are being broken into individual
characters which is cascading into the output of this script.

However, it still prints an output so I'm making a commit to be safe